Updated Opcodes and source code for August 12 2009 patch
git-svn-id: https://svn.code.sf.net/p/seq/svn/showeq/trunk@734 19a12193-7728-0410-8f54-e9f61ebc0c9f

+++ src/netstream.cpp
@@ -193,7 +193,7 @@ uint32_t NetStream::readUInt32NC()
 
 void NetStream::skipBytes(size_t byteCount)
 {
-  if ((m_lastPos - m_pos) >= byteCount)
+  if (uint32_t(m_lastPos - m_pos) >= byteCount)
   {
     m_pos += byteCount;
   }
